Trait quandl_v3::prelude::ApiParameters

pub trait ApiParameters: HasMut<ApiArguments> {
    fn api_key<S: AsRef<str>>(&mut self, api_key: S) -> &mut Self { ... }
    fn fmt(&self) -> Option<String> { ... }
}

Api parameters implemented by all queries.

Quandl API Reference

Provided Methods

fn api_key<S: AsRef<str>>(&mut self, api_key: S) -> &mut Self

Include your personal Quandl API key with your query.

fn fmt(&self) -> Option<String>

Return a string which will be appended to the query's URL given that an api key has been provided.