Trait quandl_v3::prelude::SearchParameters

pub trait SearchParameters: HasMut<SearchArguments> {
    fn query<V: AsRef<[S]>, S: AsRef<str>>(&mut self, keywords: V) -> &mut Self { ... }
    fn per_page(&mut self, n: usize) -> &mut Self { ... }
    fn page(&mut self, n: usize) -> &mut Self { ... }
    fn fmt(&self) -> Option<String> { ... }
}

Search parameters implemented by search queries.

Quandl API Reference

Provided Methods

fn query<V: AsRef<[S]>, S: AsRef<str>>(&mut self, keywords: V) -> &mut Self

Specify a vector/list of search keywords to retrieve only database/dataset related to those search terms.

fn per_page(&mut self, n: usize) -> &mut Self

Specify how many entries should be returned by search query.

fn page(&mut self, n: usize) -> &mut Self

Given there is more than one page of entries to be returned, specify which page we want to query.

fn fmt(&self) -> Option<String>

Return a string which will be appended to the query's URL given that at least one of the search parameters has been specified.